finalizado publicar artigo

+++ controllers/controllerArtigo.js
@@ -1,7 +1,27 @@
 const { Artigo, AutorArtigo, Usuario, AvaliacaoArtigo } = require('../models')
-const sequelize = require('sequelize')
+const url = require('url');
+
 const statusArtigoEnum = require('../enums/StatusArtigo')
 const tipoUsuarioEnum = require('../enums/TipoUsuario')
+
+async function validarAutorArtigo(user, idArtigo){
+    if(user.tipousuario === tipoUsuarioEnum.AUTOR){
+        const resArtigo = await AutorArtigo.findAll({
+            where: {
+                idArtigo: idArtigo,
+                idAutor: user.idUsuario
+            }
+        })
+
+        const autorArtigos = resArtigo.map( art => art.toJSON())
+
+        return autorArtigos.length > 0
+
+    }else{
+        return false
+    }
+
+}
 
 function getCreate(req, res) {
     res.render('artigo/artigoCreate', {options: 
@@ -13,6 +33,8 @@ function getCreate(req, res) {
 }
 
 async function postCreate(req,res){
+
+    if(req.session.user.tipousuario !== tipoUsuarioEnum.AUTOR) return
 
     const {artigoTitulo, artigoResumo, artigoLink, artigoAutores} = req.body
     let artigoCriado
@@ -47,6 +69,10 @@ async function postCreate(req,res){
 }
 
 async function getDelete(req, res){
+
+    const isAutorArtigo = await validarAutorArtigo(req.session.user, req.params.idArtigo)
+    if( !(isAutorArtigo || req.session.user.tipousuario === tipoUsuarioEnum.ADMIN) ) return
+
     await Artigo.destroy({ where: { idArtigo: req.params.idArtigo } }).then(
         AutorArtigo.destroy({where: { idArtigo: req.params.idArtigo }}).then(
             res.redirect('/artigo/list')
@@ -72,6 +98,14 @@ async function getUpdate(req, res){
 async function postUpdate(req, res){
 
     const {idArtigo, artigoTitulo, artigoResumo, artigoLink, artigoAutores } = req.body
+
+    console.log('entrou')
+
+    const isAutorArtigo = await validarAutorArtigo(req.session.user, idArtigo)
+
+    if(!isAutorArtigo) return
+
+    console.log('entrou 2')
 
     const resArtigo = await Artigo.update({
         titulo: artigoTitulo,
@@ -138,9 +172,10 @@ async function getAvaliador(req, res){
 
 async function postAvaliador(req, res){
 
-    console.log('entrou postavaliador')
-
     const {idArtigo, artigoAvaliadores } = req.body
+
+    if( req.session.user.tipousuario !== tipoUsuarioEnum.ADMIN ) return
+
 
     const resDeleteAvaliadorArtigo = await AvaliacaoArtigo.destroy({
         where: {idArtigo}
@@ -149,8 +184,6 @@ async function postAvaliador(req, res){
     const artigoAvaliadorInsert = artigoAvaliadores.map( (avaliador) => {
         return { idUsuario: avaliador, idArtigo: Number(idArtigo) }
     });
-
-    console.log(artigoAvaliadorInsert)
 
     const resAvaliadorArtigo = await AvaliacaoArtigo.bulkCreate(artigoAvaliadorInsert)
 
@@ -164,6 +197,32 @@ async function getPublicar(req, res){
     res.render('artigo/artigoPublicar', {
         artigos:artigos
     })
+}
+
+async function getAceitar(req, res){
+
+
+    const q = url.parse(req.url, true)
+    console.log(q.query)
+    console.log(q.query.isAceitar)
+    const isAceitar = (q.query.isAceitar === 'true' ? true : false)
+
+    if( req.session.user.tipousuario !== tipoUsuarioEnum.ADMIN ) return
+
+    const idArtigo = req.params.idArtigo
+
+    const statusArtigo = isAceitar ? statusArtigoEnum.ACEITO : statusArtigoEnum.REJEITADO
+
+    const resArtigo = await Artigo.update({
+        status: statusArtigo,
+        },
+        {
+            where: {idArtigo}
+        }
+    )
+
+    res.redirect('/artigo/publicar')
+
 }
 
 async function findByNota(){
@@ -193,32 +252,41 @@ async function findByNota(){
 
         const autores = autoresBanco.map(autoresArtigo => autoresArtigo.toJSON())
 
-        const avaliadoresArtigo = await AvaliacaoArtigo.findAll({
+        const resAvaliadoresArtigo = await AvaliacaoArtigo.findAll({
             where:{
                 idArtigo: artigo.idArtigo
             }
         })
 
-        const mediaArtigo = avaliadoresArtigo.map( avaliacao => {
-            return avaliacao.toJSON()
-        }).reduce( avaliacao => {
-            console.log('reduce')
-            console.log(avaliacao)
-            avaliacao.notaRelevancia ?? 0 * avaliacao.notaExperiencia ?? 0
-        }, 0)
 
-        console.log(mediaArtigo)
+        const avaliadoresArtigo = resAvaliadoresArtigo.map( avaliacao => {
+            return avaliacao.toJSON()
+        })
+
+        let somaNota
+        let mediaNota
+        if(avaliadoresArtigo.length > 0){
+            somaNota = avaliadoresArtigo.reduce( (acc, avaliacao) => {
+                return acc + ((Number(avaliacao.notaRelevancia) ?? 0) * (Number(avaliacao.notaExperiencia) ?? 0))
+            }, 0)
+            mediaNota = (somaNota ?? 0) / (avaliadoresArtigo.length ?? 0) ?? 0
+        } else{
+            mediaNota = 0
+        }
 
         return { 
             autores: autores, 
             nomeAutores: autores.map( autor => autor.Usuario.nome).join(', '),
             idArtigo: artigo.idArtigo,
             titulo: artigo.titulo,
-            media: mediaArtigo,
+            media: mediaNota,
         }
+
     }))
 
-    return artigos
+    const artigosSortNota = artigos.sort( (a, b) => b.media - a.media )
+
+    return artigosSortNota
 }
 
 async function findAllArtigos(){
@@ -370,5 +438,6 @@ module.exports = {
     getAvaliador,
     postAvaliador,
     getPublicar,
+    getAceitar,
     findByUsuario,
 }
